move env to be common for any contract
